Add Detail component to refactor
