+++ src/common/constants.ts
@@ -1,0 +1,2 @@
+// 추후에 Enum 혹은 상수에 빼놓기
+export const UNSELECTED_STATE = -1;
