+++ src/routes/NowPlaying.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import Header from "../components/Header";
-import { Loader, Wrapper } from "../components/common-components";
+import { Loader, Wrapper } from "../common/common-components";
 import { getNowPlaying } from "../api";
 import Movies from "../components/Movies";
